+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - [#186](https://github.com/pgerke/homebridge-freeathome-local-api/issues/186): Support SmartTouch 10" Panel device
+- [#195](https://github.com/pgerke/homebridge-freeathome-local-api/issues/195): Support a channel / device allowlist
 
 ### Changed